broke out into util files
